pyproject.toml
[project]
name = "ai-pdf-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "ebooklib>=0.20",
    "litellm>=1.82.5",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pytest>=9.0.2",
    "tenacity>=9.1.4",
]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/web/simple"
default = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
